chore: rename taskTab screens
