+++ app/tabs/taskTab/navigationConfiguration.js
@@ -2,20 +2,20 @@
 import { StackNavigator } from 'react-navigation'
 
 // Screens
-import TaskTabScreenOne from './components/TaskTabScreenOne'
-import TaskTabScreenTwo from './components/TaskTabScreenTwo'
+import TaskForm from './components/TaskForm'
+import BountyForm from './components/BountyForm'
 import TaskTabScreenThree from './components/TaskTabScreenThree'
 
 const routeConfiguration = {
-  TaskTabScreenOne: { screen: TaskTabScreenOne },
-  TaskTabScreenTwo: { screen: TaskTabScreenTwo },
+  TaskForm: { screen: TaskForm },
+  BountyForm: { screen: BountyForm },
   TaskTabScreenThree: { screen: TaskTabScreenThree },
 
 }
 // going to disable the header for now
 const stackNavigatorConfiguration = {
   headerMode: 'none',
-  initialRoute: 'TaskTabScreenOne'
+  initialRoute: 'TaskForm'
 }
 
 export const NavigatorTaskTab = StackNavigator(routeConfiguration,stackNavigatorConfiguration)
